requirements/base.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
cryptography==42.0.8
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
django-import-export==4.2.0
django-modeltranslation==0.19.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.1
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
redis==5.2.0
redis_management==1.1.0
requests==2.32.3
sqlparse==0.5.1
tablib==3.7.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
